+++ NEWS.md
@@ -1,5 +1,6 @@
 # ellmer (development version)
 
+* New `schema_df()` to describe the schema of a data frame to an LLM (#744).
 * `chat_google_gemini()` and `chat_openai_responses()` support image generation (#368).
 * `batch_*()` no longer hashes properties of the provider besides the `name`, `model`, and `base_url`. This should provide some protection from accidentally reusing the same `.json` file with different providers, while still allowing you to use the same batch file across ellmer versions.
 * `batch_*()` have a new `ignore_hash` argument that allows you to opt out of the check if you're confident the difference only arises because ellmer itself has changed.